Rename: test 폴더 구조를 main과 동일하게 변경
