--- src/test/java/hanghae/ecommerce/presentation/ProductControllerTest.java
@@ -1,7 +1,0 @@
-package hanghae.ecommerce.presentation;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class ProductControllerTest {
-}
